pip: Move project file classes into separate module
Cut and paste Python project files to a separate module (same for the
corresponding tests).

Move PyProjectTOML, SetupCFG, and SetupPY classes from main.py to a new
module project_files.py. This improves code organization and makes the
main module more focused on the core pip package manager logic. The
refactoring enhances maintainability by grouping related functionality
together and reducing the size of the main module.

Add a small summary docstring at the top of the new module.
